+++ scripts/src/form/confirmation/Text.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+const Text = ({ paymentMethodInfo }) => {
+  return <div>{paymentMethodInfo.text}</div>;
+};
+
+export default Text;