+++ scripts/src/form/confirmation/payment.tsx
@@ -4,6 +4,7 @@ import PaypalClassic from "./PaypalClassic";
 import CCAvenue from "./CCAvenue";
 import ManualApproval from "./ManualApproval";
 import Redirect from "./Redirect";
+import Text from "./Text";
 import sanitize from "../../sanitize";
 import ExpressionParser from "../../common/ExpressionParser";
 import { IPaymentProps, IPaymentInfo } from "../interfaces";
@@ -13,8 +14,12 @@ let Components = {
   ccavenue: CCAvenue,
   manual_approval: ManualApproval,
   manual_approval_2: ManualApproval,
-  redirect: Redirect
+  redirect: Redirect,
+  text: Text
 };
+
+// Payment methods that are always shown, even if the user has fully paid.
+const PAYMENT_METHODS_ALWAYS_SHOWN = ["redirect", "text"];
 
 class Payment extends React.Component<IPaymentProps, any> {
   constructor(props: any) {
@@ -30,7 +35,11 @@ class Payment extends React.Component<IPaymentProps, any> {
     return paymentMethods.map(paymentMethod => {
       // If fullyPaid is true, only gets payment methods that apply regardless of whether
       // one has fully paid (in this case, the "redirect" payment method).
-      if (fullyPaid && paymentMethod !== "redirect") return;
+      if (
+        fullyPaid &&
+        PAYMENT_METHODS_ALWAYS_SHOWN.indexOf(paymentMethod) === -1
+      )
+        return;
 
       var MyComponent = Components[paymentMethod];
       if (!MyComponent) return;
@@ -101,6 +110,18 @@ class Payment extends React.Component<IPaymentProps, any> {
   render() {
     if (!this.props.paymentMethods) {
       return "";
+    }
+    if (
+      (!this.props.paymentInfo_owed ||
+        this.props.paymentInfo_owed.total === 0) &&
+      (!this.props.paymentInfo ||
+        !this.props.paymentInfo.items ||
+        this.props.paymentInfo.items.length === 0) &&
+      (!this.props.paymentInfo_received ||
+        this.props.paymentInfo_owed.total === 0)
+    ) {
+      // No payment involved, so just render manual payment buttons.
+      return <>{this.getPaymentMethods(false)}</>;
     }
     return (
       <div>